+++ modules/Test/routes.php
@@ -1,0 +1,3 @@
+<?php
+/** @var \Core\Router $router */
+$router->get('/test', [\Modules\Test\Http\Controllers\TestHomeController::class, 'index']);